Allow fractional M and K specifications

+++ svg_prime/models/conv_lstm.py
@@ -12,10 +12,10 @@ class ConvLSTM(nn.Module):
         self.batch_size = batch_size
         self.n_layers = n_layers
         self.image_size = image_size
-        self.embed = nn.Conv2d(input_size, hidden_size*expand, 3, 1, 1) # Initial embedding conv layer
-        self.lstm = nn.ModuleList([ConvLSTMCell(hidden_size*expand, hidden_size*expand, (3, 3), True) for i in range(self.n_layers)])
+        self.embed = nn.Conv2d(input_size, int(hidden_size*expand), 3, 1, 1) # Initial embedding conv layer
+        self.lstm = nn.ModuleList([ConvLSTMCell(int(hidden_size*expand), int(hidden_size*expand), (3, 3), True) for i in range(self.n_layers)])
         self.output = nn.Sequential(
-                nn.Conv2d(hidden_size*expand, output_size, 3, 1, 1),
+                nn.Conv2d(int(hidden_size*expand), output_size, 3, 1, 1),
                 #nn.BatchNorm1d(output_size),
                 nn.Tanh())
         self.hidden = self.init_hidden()
@@ -44,10 +44,10 @@ class ConvGaussianLSTM(nn.Module):
         self.image_size = image_size
         self.n_layers = n_layers
         self.batch_size = batch_size
-        self.embed = nn.Conv2d(input_size, hidden_size*expand, 3, 1, 1)
-        self.lstm = nn.ModuleList([ConvLSTMCell(hidden_size*expand, hidden_size*expand, (3, 3), True) for _ in range(self.n_layers)])
-        self.mu_net = nn.Conv2d(hidden_size*expand, output_size, 3, 1, 1)
-        self.logvar_net = nn.Conv2d(hidden_size*expand, output_size, 3, 1, 1)
+        self.embed = nn.Conv2d(input_size, int(hidden_size*expand), 3, 1, 1)
+        self.lstm = nn.ModuleList([ConvLSTMCell(int(hidden_size*expand), int(hidden_size*expand), (3, 3), True) for _ in range(self.n_layers)])
+        self.mu_net = nn.Conv2d(int(hidden_size*expand), output_size, 3, 1, 1)
+        self.logvar_net = nn.Conv2d(int(hidden_size*expand), output_size, 3, 1, 1)
         self.hidden = self.init_hidden()
 
     def init_hidden(self):
